log_collectors/nginx_logger/requirements.txt
iniconfig==2.0.0
packaging==23.1
pika==1.3.2
pluggy==1.0.0
pytest==7.3.1
python-dotenv==1.0.0
